Feature/#74

+++ app/src/main/java/com/d_vide/D_VIDE/app/data/remote/responseDTO/ReviewIdDTO.kt
@@ -1,5 +1,5 @@
 package com.d_vide.D_VIDE.app.data.remote.responseDTO
 
 data class ReviewIdDTO(
-    val reviewId: Int
+    val reviewId: Long
 )
